Moved completed conversion functions to separate files
