+++ ls/ls_formatter.go
@@ -1,0 +1,191 @@
+package ls
+
+import (
+	"github.com/arduino/go-paths-helper"
+	"go.bug.st/lsp"
+	"go.bug.st/lsp/jsonrpc"
+)
+
+func (ls *INOLanguageServer) createClangdFormatterConfig(logger jsonrpc.FunctionLogger, cppuri lsp.DocumentURI) (func(), error) {
+	// clangd looks for a .clang-format configuration file on the same directory
+	// pointed by the uri passed in the lsp command parameters.
+	// https://github.com/llvm/llvm-project/blob/64d06ed9c9e0389cd27545d2f6e20455a91d89b1/clang-tools-extra/clangd/ClangdLSPServer.cpp#L856-L868
+	// https://github.com/llvm/llvm-project/blob/64d06ed9c9e0389cd27545d2f6e20455a91d89b1/clang-tools-extra/clangd/ClangdServer.cpp#L402-L404
+
+	config := `# See: https://releases.llvm.org/11.0.1/tools/clang/docs/ClangFormatStyleOptions.html
+---
+Language: Cpp
+# LLVM is the default style setting, used when a configuration option is not set here
+BasedOnStyle: LLVM
+AccessModifierOffset: -2
+AlignAfterOpenBracket: Align
+AlignConsecutiveAssignments: false
+AlignConsecutiveBitFields: false
+AlignConsecutiveDeclarations: false
+AlignConsecutiveMacros: false
+AlignEscapedNewlines: DontAlign
+AlignOperands: Align
+AlignTrailingComments: true
+AllowAllArgumentsOnNextLine: true
+AllowAllConstructorInitializersOnNextLine: true
+AllowAllParametersOfDeclarationOnNextLine: true
+AllowShortBlocksOnASingleLine: Always
+AllowShortCaseLabelsOnASingleLine: true
+AllowShortEnumsOnASingleLine: true
+AllowShortFunctionsOnASingleLine: Empty
+AllowShortIfStatementsOnASingleLine: Always
+AllowShortLambdasOnASingleLine: Empty
+AllowShortLoopsOnASingleLine: true
+AlwaysBreakAfterDefinitionReturnType: None
+AlwaysBreakAfterReturnType: None
+AlwaysBreakBeforeMultilineStrings: false
+AlwaysBreakTemplateDeclarations: No
+BinPackArguments: true
+BinPackParameters: true
+# Only used when "BreakBeforeBraces" set to "Custom"
+BraceWrapping:
+  AfterCaseLabel: false
+  AfterClass: false
+  AfterControlStatement: Never
+  AfterEnum: false
+  AfterFunction: false
+  AfterNamespace: false
+  #AfterObjCDeclaration:
+  AfterStruct: false
+  AfterUnion: false
+  AfterExternBlock: false
+  BeforeCatch: false
+  BeforeElse: false
+  BeforeLambdaBody: false
+  BeforeWhile: false
+  IndentBraces: false
+  SplitEmptyFunction: false
+  SplitEmptyRecord: false
+  SplitEmptyNamespace: false
+# Java-specific
+#BreakAfterJavaFieldAnnotations:
+BreakBeforeBinaryOperators: NonAssignment
+BreakBeforeBraces: Attach
+BreakBeforeTernaryOperators: true
+BreakConstructorInitializers: BeforeColon
+BreakInheritanceList: BeforeColon
+BreakStringLiterals: false
+ColumnLimit: 0
+# "" matches none
+CommentPragmas: ""
+CompactNamespaces: false
+ConstructorInitializerAllOnOneLineOrOnePerLine: true
+ConstructorInitializerIndentWidth: 2
+ContinuationIndentWidth: 2
+Cpp11BracedListStyle: false
+DeriveLineEnding: true
+DerivePointerAlignment: true
+DisableFormat: false
+# Docs say "Do not use this in config files". The default (LLVM 11.0.1) is "false".
+#ExperimentalAutoDetectBinPacking:
+FixNamespaceComments: false
+ForEachMacros: []
+IncludeBlocks: Preserve
+IncludeCategories: []
+# "" matches none
+IncludeIsMainRegex: ""
+IncludeIsMainSourceRegex: ""
+IndentCaseBlocks: true
+IndentCaseLabels: true
+IndentExternBlock: Indent
+IndentGotoLabels: false
+IndentPPDirectives: None
+IndentWidth: 2
+IndentWrappedFunctionNames: false
+InsertTrailingCommas: None
+# Java-specific
+#JavaImportGroups:
+# JavaScript-specific
+#JavaScriptQuotes:
+#JavaScriptWrapImports
+KeepEmptyLinesAtTheStartOfBlocks: true
+MacroBlockBegin: ""
+MacroBlockEnd: ""
+# Set to a large number to effectively disable
+MaxEmptyLinesToKeep: 100000
+NamespaceIndentation: None
+NamespaceMacros: []
+# Objective C-specific
+#ObjCBinPackProtocolList:
+#ObjCBlockIndentWidth:
+#ObjCBreakBeforeNestedBlockParam:
+#ObjCSpaceAfterProperty:
+#ObjCSpaceBeforeProtocolList
+PenaltyBreakAssignment: 1
+PenaltyBreakBeforeFirstCallParameter: 1
+PenaltyBreakComment: 1
+PenaltyBreakFirstLessLess: 1
+PenaltyBreakString: 1
+PenaltyBreakTemplateDeclaration: 1
+PenaltyExcessCharacter: 1
+PenaltyReturnTypeOnItsOwnLine: 1
+# Used as a fallback if alignment style can't be detected from code (DerivePointerAlignment: true)
+PointerAlignment: Right
+RawStringFormats: []
+ReflowComments: false
+SortIncludes: false
+SortUsingDeclarations: false
+SpaceAfterCStyleCast: false
+SpaceAfterLogicalNot: false
+SpaceAfterTemplateKeyword: false
+SpaceBeforeAssignmentOperators: true
+SpaceBeforeCpp11BracedList: false
+SpaceBeforeCtorInitializerColon: true
+SpaceBeforeInheritanceColon: true
+SpaceBeforeParens: ControlStatements
+SpaceBeforeRangeBasedForLoopColon: true
+SpaceBeforeSquareBrackets: false
+SpaceInEmptyBlock: false
+SpaceInEmptyParentheses: false
+SpacesBeforeTrailingComments: 2
+SpacesInAngles: false
+SpacesInCStyleCastParentheses: false
+SpacesInConditionalStatement: false
+SpacesInContainerLiterals: false
+SpacesInParentheses: false
+SpacesInSquareBrackets: false
+Standard: Auto
+StatementMacros: []
+TabWidth: 2
+TypenameMacros: []
+# Default to LF if line endings can't be detected from the content (DeriveLineEnding).
+UseCRLF: false
+UseTab: Never
+WhitespaceSensitiveMacros: []
+`
+	try := func(conf *paths.Path) bool {
+		if c, err := conf.ReadFile(); err != nil {
+			logger.Logf("    error reading custom formatter config file %s: %s", conf, err)
+		} else {
+			logger.Logf("    using custom formatter config file %s", conf)
+			config = string(c)
+		}
+		return true
+	}
+
+	if sketchFormatterConf := ls.sketchRoot.Join(".clang-format"); sketchFormatterConf.Exist() {
+		// If a custom config is present in the sketch folder, use that one
+		try(sketchFormatterConf)
+	} else if ls.config.FormatterConf != nil && ls.config.FormatterConf.Exist() {
+		// Otherwise if a global config file is present, use that one
+		try(ls.config.FormatterConf)
+	}
+
+	targetFile := cppuri.AsPath()
+	if targetFile.IsNotDir() {
+		targetFile = targetFile.Parent()
+	}
+	targetFile = targetFile.Join(".clang-format")
+	cleanup := func() {
+		targetFile.Remove()
+		logger.Logf("    formatter config cleaned")
+	}
+	logger.Logf("    writing formatter config in: %s", targetFile)
+	err := targetFile.WriteFile([]byte(config))
+	return cleanup, err
+}
